fix: remove bad geocoding api call
There was an api call that added the address into the title bar of each car park entry on the map. Since getting this working would require me to unlock my api key to use a specific api, I decided to give each window a generic title instead, so I can keep the api key as restricted as possible.